lydian: Add noop stop() on Mock clients.
Change adds noop stop() method on dummy client and writers. Mock clients
are added to have resiliency in lydian so that when some of the clients
are not available at enpoints , rest of the application doesn't stop
functioning.

In the end, stop() method is called to close all the recorders. In case
of mock clients, w/o stop() method AttributeError would come. This
change fixes that.

Change-Id: I9c6a9f845eeb4a4d6ea18a19ecc69a34e8756999